requirements.txt
PyQt6>=6.5.0
SQLAlchemy>=2.0.0
psycopg2
reportlab>=4.0.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
matplotlib>=3.7.0
requests>=2.31.0
bcrypt>=4.0.0
